Starts WPILIB Data Logging (finished)
To find the logs, open "Data Log Tool" on your driverstation (downloaded with other wpilib applications such as smartdashboard). Go to download tab in the bottom left and set team number/address to 10.8.62.2 (address of RIO), set username to lvuser, and keep password blank. If the connection is made after logs are created you can just hit the download button and select a file path. Load the file onto AdvantageScope and just select which graphs you want to view. not tested on an actual robot yet but have used a spare roborio and it worked (gave the file) although it gave no values.